Support MobileStorageAdapter
